use a new executor for every started cluster instance
this executor handles the background thread that checks if the server is ready and the thread forwarding stdout to the test process which started the test-server
reusing it after the first instance was stopped, lead to RejectedExecutionEceptions as the pool was shut down. In effect we were not able to run more than to tests in a row,
reusing the same static instance of the test-server.